pyproject.toml
[tool.poetry]
name = "bigdatacorp-wrapper"
version = "0.1.0"
description = ""
authors = ["Ricardo Gaya <rrgaya@gmail.com>", "Ricardo Gaya <ricardo.gaya@linker.com.br>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
